+++ tests/sdk/python/python_test.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2020, Pulumi Corporation.
+// Copyright 2016-2021, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,6 +19,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"regexp"
 	"sort"
 	"strings"
 	"testing"
@@ -360,6 +361,34 @@ func TestHelmLocal(t *testing.T) {
 	}
 	options := baseOptions.With(integration.ProgramTestOptions{
 		Dir: filepath.Join(cwd, "helm-local", "step1"),
+		ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
+			// Verify resource creation order using the Event stream. The Chart resources must be created
+			// first, followed by the dependent ConfigMap. (The ConfigMap doesn't actually need the Chart, but
+			// it creates almost instantly, so it's a good choice to test creation ordering)
+			cmRegex := regexp.MustCompile(`ConfigMap::nginx-server-block`)
+			svcRegex := regexp.MustCompile(`Service::nginx`)
+			deployRegex := regexp.MustCompile(`Deployment::nginx`)
+			dependentRegex := regexp.MustCompile(`ConfigMap::foo`)
+
+			var configmapFound, serviceFound, deploymentFound, dependentFound bool
+			for _, e := range stackInfo.Events {
+				if e.ResOutputsEvent != nil {
+					switch {
+					case cmRegex.MatchString(e.ResOutputsEvent.Metadata.URN):
+						configmapFound = true
+					case svcRegex.MatchString(e.ResOutputsEvent.Metadata.URN):
+						serviceFound = true
+					case deployRegex.MatchString(e.ResOutputsEvent.Metadata.URN):
+						deploymentFound = true
+					case dependentRegex.MatchString(e.ResOutputsEvent.Metadata.URN):
+						dependentFound = true
+					}
+					assert.Falsef(t, dependentFound && !(configmapFound && serviceFound && deploymentFound),
+						"dependent ConfigMap created before all chart resources were ready")
+					fmt.Println(e.ResOutputsEvent.Metadata.URN)
+				}
+			}
+		},
 		EditDirs: []integration.EditDir{
 			{
 				Dir:             filepath.Join(cwd, "helm-local", "step2"),